requirements.txt
Flask
pyjwt
pymongo
flask-restful
flask-pymongo
h5py
numpy
matplotlib
gunicorn
